moved private key to hardhat helpers
