+++ ethwrite/src/index.ts
@@ -1,7 +1,7 @@
-import { expose, util } from '@chainlink/ea-bootstrap'
+import { expose } from '@chainlink/ea-bootstrap'
 import { makeExecute } from './adapter'
 import { makeConfig } from './config'
 
 const NAME = 'ETHWRITE'
 
-export = { NAME, makeExecute, makeConfig, ...expose(util.wrapExecute(makeExecute())) }
+export = { NAME, makeExecute, makeConfig, ...expose(makeExecute()) }
