+++ hardhat.config.js
@@ -3,6 +3,7 @@
 
 // const { MNEMONIC } = require('./test.env')
 // const MNEMONIC = 'novel mobile inform nurse circle spoon cricket soup crowd clip hawk glad'
+const { TESTING_PRIVATE_KEY } = require('./test-helpers/src/hardhat')
 
 module.exports = {
   defaultNetwork: 'hardhat',
@@ -11,9 +12,7 @@ module.exports = {
       chainId: 31337,
       accounts: [
         {
-          privateKey:
-            process.env.PRIVATE_KEY ||
-            '0x90125e49d93a24cc8409d1e00cc69c88919c6826d8bbabb6f2e1dc8213809f4c',
+          privateKey: TESTING_PRIVATE_KEY,
           balance: '10000000000000000000000',
         },
       ],
